chore(client): dashboard
- renamed files from broadpage to dashboard
- modified template design
- refactored code in dashboard
