--- template/broadcast.html
@@ -1,153 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-    <!--Import Google Icon Font-->
-      <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-    <!--Import materialize.css-->
-    <link type="text/css" rel="stylesheet" href="css/materialize.min.css"  media="screen,projection"/>
-    <!--Import Css-->
-    <link type="text/css" rel="stylesheet" href="style.css">
-
-  </head>
-  <body>
-    
-
-    <nav>
-      <div class="nav-wrapper green darken-4">
-        <a href="#" class="brand-logo center"><img width="70px" src="image/postitL.png" alt="PostIt Logo"></a>
-        <ul id="nav-mobile" class="right">
-          <li><a class="modal-trigger" href="#modal1"><i href="#modal1" class="modal-trigger white-text material-icons">settings</i></a></li>
-        </ul>
-      </div>
-    </nav>
-    
-    <div id="modal1" class="modal">
-      <div class="modal-content">
-        <h4>User Profile</h4>
-        <form class="">
-          <div class="row">
-            <div class="input-field col s12 m12">
-              <input class="validate" type="text" id="userName" data-error="wrong" data-success="right">
-              <label for="userName">Username</label>
-            </div>
-          </div>
-          <div class="row">
-            <div class="input-field col s12 m12">
-              <input class="validate" type="email" id="email" data-error="wrong" data-success="right">
-              <label for="email">Email</label>
-            </div>
-          </div>
-          <div class="row">
-            <div class="input-field col s12 m12">
-              <input class="validate" type="text" id="phoneNumber" data-error="wrong" data-success="right">
-              <label for="phoneNumber">Phone Number</label>
-            </div>
-          </div>
-        </form>
-      </div>
-      <div class="modal-footer">
-        <button class="btn-large green darken-4 waves-effect" type="submit" name="action">
-          <a class="white-text" href="broadcast.html">Submit</a>
-        </button>
-      </div>
-    </div>
-          
-    <div class="container">
-      <br>
-      <br>
-      <div class="row center">
-        <div>
-          <a class="" href="group.html">
-            <img width="70px" src="image/gIcon.png">
-            <p>Create Group</p>
-          </a>
-        </div>
-      </div>
-
-      <div class="row center">
-        <div class="input-field">
-          <i class="material-icons prefix">search</i>
-          <input id="icon_prefix" type="text">
-          <label for="icon_prefix">Search Groups</label>
-        </div>
-      </div>
-
-      <div class="row">
-        <div class="col s12 m12">
-          <div class="card green darken-1">
-            <div class="card-content white-text">
-              <span class="card-title" style="font-weight:bold">Sample Group 1</span>
-              <p>I am a very simple card. I am good at containing small bits of information of the very last message posted in the group.
-              I am convenient because I require little markup to use effectively.</p>
-              <hr>
-              <p>I am a very simple card. I am good at containing small bits of information of the very last message posted in the group.
-              I am convenient because I require little markup to use effectively.</p>
-              <hr>
-              <p>I am a very simple card. I am good at containing small bits of information of the very last message posted in the group.
-              I am convenient because I require little markup to use effectively.</p>
-              <hr>
-            </div>
-            <div class="card-action">
-              <a class="white-text" href="messageBoard.html">Go to Group</a>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      <div class="row">
-        <div class="col s12 m12">
-          <div class="card green darken-1">
-            <div class="card-content white-text">
-              <span class="card-title" style="font-weight:bold">Sample Group 2</span>
-              <p>I am a very simple card. I am good at containing small bits of information of the very last message posted in the group.
-              I am convenient because I require little markup to use effectively.</p>
-            </div>
-            <div class="card-action">
-              <a class="white-text" href="messageBoard.html">Go to Group</a>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      <div class="row">
-        <div class="col s12 m12">
-          <div class="card green darken-1">
-            <div class="card-content white-text">
-              <span class="card-title" style="font-weight:bold">Sample Group 3</span>
-              <p>I am a very simple card. I am good at containing small bits of information of the very last message posted in the group.
-              I am convenient because I require little markup to use effectively.</p>
-              <hr>
-              <p>I am a very simple card. I am good at containing small bits of information of the very last message posted in the group.
-              I am convenient because I require little markup to use effectively.</p>
-              <hr>
-            </div>
-            <div class="card-action">
-              <a class="white-text" href="messageBoard.html">Go to Group</a>
-            </div>
-          </div>
-        </div>
-      </div>
-
-
-    </div>
-
-
-    
-            
-
-
-    <!-- jQuery first, then Tether, then Bootstrap JS. -->
-    <script type="text/javascript" src="https://code.jquery.com/jquery-3.2.1.min.js"></script>
-    <script type="text/javascript" src="js/materialize.min.js"></script>
-
-    <script type="text/javascript">
-      $('document').ready(function(){
-      $('.modal').modal();
-  });
-    </script>
-  </body>
-</html>
