+++ template/login.html
@@ -18,12 +18,13 @@
 
     <div class="container">
       <div class="row centerInDiv">
-        <img width="30%" src="image/postitD.png" alt="PostIt Logo">
+        <img width="30%" style="padding-top: 15vh" src="image/dark_logo" alt="PostIt Logo">
       </div>
 
       <div class="row">
 
-        <div class="container col s12 m12">
+        <div class="col s12 m1"></div>
+        <div class="container col s12 m10">
 
         <form class="">
           <div class="row">
@@ -42,7 +43,7 @@
             <a class="black-text" href="#">Forgot Password?</a>
           </div>
           <div class="row centerInDiv">
-            <button class="btn-large green darken-4 waves-effect" type="submit" name="action"><a class="white-text" href="broadcast.html">Sign In</a>
+            <button class="btn-large green darken-4 waves-effect" type="submit" name="action"><a class="white-text" href="dashboard.html">Sign In</a>
               <i class="material-icons right">send</i>
             </button>
           </div>
