+++ template/style.css
@@ -1,3 +1,7 @@
+#navbar {
+  background-color: #002102;
+}
+
 .noLinkDeco{
 
   text-decoration: none;
